+++ PSPDFKit.embeddedframework/PSPDFKit.framework/Versions/A/Headers/PSPDFFlexibleAnnotationToolbar.h
@@ -95,3 +95,11 @@ typedef UIImage *(^PSPDFAnnotationGroupItemConfigurationBlock)(PSPDFAnnotationGr
 + (instancetype)itemWithType:(NSString *)type variant:(NSString *)variant configurationBlock:(PSPDFAnnotationGroupItemConfigurationBlock)block;
 
 @end
+
+@interface PSPDFFlexibleAnnotationToolbar (SubclassingHooks)
+
+- (UIButton *)doneButton;
+- (UIButton *)undoButton;
+- (UIButton *)redoButton;
+
+@end