+++ PSPDFKit.embeddedframework/PSPDFKit.framework/Versions/A/Headers/PSPDFIconFit.h
@@ -50,3 +50,6 @@ typedef NS_ENUM(NSUInteger, PSPDFIconFitScaleType) {
 - (NSString *)PDFString;
 
 @end
+
+extern NSString *const PSPDFIconFitScaleModeTransformer;
+extern NSString *const PSPDFIconFitScaleTypeTransformer;