+++ PSPDFKit.embeddedframework/PSPDFKit.framework/Versions/A/Headers/PSPDFSoundAnnotationView.h
@@ -52,4 +52,4 @@
 @property (nonatomic, strong, readonly) UILabel *totalTimeLabel;
 @property (nonatomic, strong, readonly) UILabel *statusLabel;
 @property (nonatomic, strong, readonly) UISlider *progressSlider;
-@end;
+@end